requirements.txt
streamlit
google-genai
openai-whisper
gTTS
python-dotenv